// projects/apppensionista/src/app/features/actualizar-datos/models/formulario.model.ts
export class FormularioModel{
    TipoDocumentoId: number = 0;
    NumeroDocumento: string = "";
    Correo: string = "";
    TelefonoFijo: string = "";
    Celular: string = "";
    CodCelularPais: string = "PE";
    Direccion: string = "";
    UbigeoDepartamento: string = "";
    UbigeoProvincia: string = "";
    UbigeoDistrito: string = "";
    ReferenciaDireccion: string = "";
    EsAntiguo: boolean = false;
}